+++ packages/react/src/hooks/codegen/createUseWriteContract.ts
@@ -290,5 +290,8 @@ type Variables<
             | undefined
         }
       : Compute<ChainIdParameter<config, chainId>>) &
-    ConnectorParameter & { __mode?: 'prepared' }
+    ConnectorParameter & {
+      /** @deprecated */
+      __mode?: 'prepared'
+    }
 >