+++ packages/react/src/hooks/useWriteContract.test-d.ts
@@ -20,7 +20,6 @@ test('context', () => {
     mutation: {
       onMutate(variables) {
         expectTypeOf(variables).toMatchTypeOf<{
-          __mode?: 'prepared'
           chainId?: number | undefined
           abi: Abi
           functionName: string
@@ -33,7 +32,6 @@ test('context', () => {
         expectTypeOf(context).toEqualTypeOf<typeof contextValue | undefined>()
 
         expectTypeOf(variables).toMatchTypeOf<{
-          __mode?: 'prepared'
           chainId?: number | undefined
           abi: Abi
           functionName: string
@@ -45,7 +43,6 @@ test('context', () => {
         expectTypeOf(context).toEqualTypeOf<typeof contextValue>()
 
         expectTypeOf(variables).toMatchTypeOf<{
-          __mode?: 'prepared'
           chainId?: number | undefined
           abi: Abi
           functionName: string
@@ -58,7 +55,6 @@ test('context', () => {
         expectTypeOf(context).toEqualTypeOf<typeof contextValue | undefined>()
 
         expectTypeOf(variables).toMatchTypeOf<{
-          __mode?: 'prepared'
           chainId?: number | undefined
           abi: Abi
           functionName: string
@@ -89,7 +85,6 @@ test('context', () => {
         expectTypeOf(context).toEqualTypeOf<typeof contextValue | undefined>()
 
         expectTypeOf(variables).toMatchTypeOf<{
-          __mode?: 'prepared'
           chainId?: number | undefined
           abi: typeof abi.erc20
           functionName: 'transferFrom'
@@ -105,7 +100,6 @@ test('context', () => {
           readonly [Address, Address, bigint]
         >()
         expectTypeOf(variables).toMatchTypeOf<{
-          __mode?: 'prepared'
           chainId?: number | undefined
           abi: typeof abi.erc20
           functionName: 'transferFrom'
@@ -118,7 +112,6 @@ test('context', () => {
         expectTypeOf(context).toEqualTypeOf<typeof contextValue | undefined>()
 
         expectTypeOf(variables).toMatchTypeOf<{
-          __mode?: 'prepared'
           chainId?: number | undefined
           abi: typeof abi.erc20
           functionName: 'transferFrom'